+++ src/main/java/service/CompanyDirectoryService.java
@@ -1,9 +1,17 @@
 package service;
 
+import Scanner.Draft_ReadCSVFile_BufferedReader;
+import storage.Person;
+
 public class CompanyDirectoryService implements ICompanyDirectoryService {
 
     public void searchByName(String Name) {
         //Заглушка для поиска
+
+        Person person = Person.newBuilderPerson().build();
+
+
+
         String FIO = "Иванов ИванИванович";
         FIO = FIO.toLowerCase();
         Name = Name.toLowerCase();
